requirements.txt
thordata-sdk>=1.3.0
python-dotenv
pandas
colorama
requests